+++ src/tiff/mod.rs
@@ -8,7 +8,8 @@
 //!
 
 pub use self::decoder::TIFFDecoder;
-pub use self::decoder::ByteOrder;
+pub use self::stream::ByteOrder;
 
 mod decoder;
 mod ifd;
+mod stream;
